OS-7.8: added role edit dropDown list in userEditView rename DataAccess folder
