+++ OnlineStore_Epam2018/OnlineStore_Epam2018/Views/User/Details.cshtml
@@ -55,7 +55,7 @@
         </dt>
 
         <dd>
-            @Html.DisplayFor(model => model.RoleName)
+            @Html.DisplayFor(model => model.Role.Name)
         </dd>
 
         <dt>
