+++ OnlineStore_Epam2018/OnlineStore_Epam2018/Views/User/EditUserAutorithationInfo.cshtml
@@ -12,7 +12,7 @@
     @Html.HiddenFor(model => model.UserId);
     @Html.HiddenFor(model => model.Login);
     @Html.HiddenFor(model => model.Password);
-    @Html.HiddenFor(model => model.RoleName);
+    @Html.HiddenFor(model => model.Role.RoleId);
     @Html.HiddenFor(model => model.UserId);
 
     <div class="form-horizontal">
