+++ OnlineStore_Epam2018/OnlineStore_Epam2018/Web.config
@@ -55,12 +55,12 @@
   </system.codedom>
   
     <system.webServer>
-      <httpErrors errorMode="Custom" existingResponse="Replace">
+      <!--<httpErrors errorMode="Custom" existingResponse="Replace">
       <clear/>
       <error statusCode="404" path="/Error/NotFound" responseMode="ExecuteURL"/>
       <error statusCode="401" path="/Login/Entrance" responseMode="ExecuteURL"/>
       <error statusCode="500" path="/Error/ServerError" responseMode="ExecuteURL"/>
-    </httpErrors>
+    </httpErrors>-->
   </system.webServer>
 
   <system.serviceModel>
