+++ OnlineStore_Epam2018/SA.OnlineStore.Bussines/Service/Implementation/OrderService.cs
@@ -4,7 +4,7 @@
     using SA.OnlineStore.Bussines.Service;
     using SA.OnlineStore.Common.Entity;
     using SA.OnlineStore.DataAccess.Implements;
-    using SA.OnlineStore.DataAccess.Service;
+    using SA.OnlineStore.DataAccess.Repositorys;
     using System;
     using System.Collections.Generic;
     using System.Linq;
