+++ OnlineStore_Epam2018/SA.OnlineStore.Bussines/Service/Implementation/ProductService.cs
@@ -5,7 +5,7 @@
     using SA.OnlineStore.Common.Cache;
     using SA.OnlineStore.Common.Entity;
     using SA.OnlineStore.DataAccess.Implements;
-    using SA.OnlineStore.DataAccess.Service;
+    using SA.OnlineStore.DataAccess.Repositorys;
     using System.Collections.Generic;
     using System.Linq;
     #endregion
