+++ OnlineStore_Epam2018/SA.OnlineStore.Bussines/Service/Implementation/UserService.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using SA.OnlineStore.Common.Entity;
 using SA.OnlineStore.DataAccess.Implements;
-using SA.OnlineStore.DataAccess.Service;
+using SA.OnlineStore.DataAccess.Repositorys;
 
 namespace SA.OnlineStore.Bussines.Service.Implementation
 {
