+++ OnlineStore_Epam2018/SA.OnlineStore.DataAccess/Repositorys/IProductRepository.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SA.OnlineStore.DataAccess.Service
+namespace SA.OnlineStore.DataAccess.Repositorys
 {
     public interface IProductRepository : IRepository<Product>
     {
