+++ OnlineStore_Epam2018/SA.OnlineStore.DataAccess/Repositorys/IRepository.cs
