+++ OnlineStore_Epam2018/SA.OnlineStore.DataAccess/Repositorys/Implementation/RoleRepository.cs
@@ -10,7 +10,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SA.OnlineStore.DataAccess.Service.Implementation
+namespace SA.OnlineStore.DataAccess.Repositorys.Implementation
 {
     public class RoleRepository : IRepository<Role>
     {
