+++ OnlineStore_Epam2018/SA.OnlineStore.DataAccess/SA.OnlineStore.DataAccess.csproj
@@ -71,24 +71,24 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Reference.svcmap</DependentUpon>
     </Compile>
-    <Compile Include="Service\IOrderRepository.cs" />
-    <Compile Include="Service\IProductRepository.cs" />
-    <Compile Include="Service\IRepository.cs" />
+    <Compile Include="Repositorys\IOrderRepository.cs" />
+    <Compile Include="Repositorys\IProductRepository.cs" />
+    <Compile Include="Repositorys\IRepository.cs" />
     <Compile Include="Implements\RealizationImplementation.cs" />
-    <Compile Include="Service\Implementation\BasketRepository.cs" />
-    <Compile Include="Service\Implementation\CategoryRepository.cs" />
-    <Compile Include="Service\Implementation\EmailRepository.cs" />
-    <Compile Include="Service\Implementation\OrderRepository.cs" />
-    <Compile Include="Service\Implementation\PhoneRepository.cs" />
-    <Compile Include="Service\Implementation\ProductRepository.cs" />
-    <Compile Include="Service\Implementation\PublicityRepository.cs" />
-    <Compile Include="Service\Implementation\RoleRepository.cs" />
-    <Compile Include="Service\Implementation\SeasonRepository.cs" />
+    <Compile Include="Repositorys\Implementation\BasketRepository.cs" />
+    <Compile Include="Repositorys\Implementation\CategoryRepository.cs" />
+    <Compile Include="Repositorys\Implementation\EmailRepository.cs" />
+    <Compile Include="Repositorys\Implementation\OrderRepository.cs" />
+    <Compile Include="Repositorys\Implementation\PhoneRepository.cs" />
+    <Compile Include="Repositorys\Implementation\ProductRepository.cs" />
+    <Compile Include="Repositorys\Implementation\PublicityRepository.cs" />
+    <Compile Include="Repositorys\Implementation\RoleRepository.cs" />
+    <Compile Include="Repositorys\Implementation\SeasonRepository.cs" />
     <Compile Include="Dependency\DataAccessDependencyRegistry.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Service\Implementation\UserRepository.cs" />
-    <Compile Include="Service\IRealizationImplementation.cs" />
-    <Compile Include="Service\IUserRepository.cs" />
+    <Compile Include="Repositorys\Implementation\UserRepository.cs" />
+    <Compile Include="Repositorys\IRealizationImplementation.cs" />
+    <Compile Include="Repositorys\IUserRepository.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config">
@@ -114,7 +114,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Include="packages.config" />
-    <None Include="Service\packages.config" />
+    <None Include="Repositorys\packages.config" />
   </ItemGroup>
   <ItemGroup>
     <WCFMetadata Include="Connected Services\" />
